requirements.txt
dwave-ocean-sdk>=4.5.0
matplotlib==3.5.2
descartes==1.1.0
pyshp==2.3.0
pandas>=1.3.5
Shapely==1.8.2